requirements.txt
pyrogram==2.0.106
tgcrypto>=1.2.3
lyricsgenius>=3.0.0,<4.0
PyYAML>=6.0